Remove Kafka Connect properties from root of KafkaMirrorMaker2Spec
This commit refactors KafkaMirrorMaker2Spec to extend a new
AbstractKafkaConnectSpec that doesn't include the properties that
should only apply at the `spec.clusters` level. It also uses the
AbstractKafkaConnectSpec as the super class of the KafkaConnectSpec.
