+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaMirrorMaker2Spec.java
@@ -4,9 +4,7 @@
  */
 package io.strimzi.api.kafka.model;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -22,18 +20,17 @@ import lombok.EqualsAndHashCode;
         builderPackage = "io.fabric8.kubernetes.api.builder"
 )
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({"replicas", "connectCluster", "config", "image",
+@JsonPropertyOrder({"replicas", "connectCluster", "image",
         "livenessProbe", "readinessProbe", "jvmOptions",
         "affinity", "tolerations", "logging", "metrics", "tracing", "template", "clusters", "mirrors"})
 @EqualsAndHashCode(callSuper = true, doNotUseGetters = true)
-public class KafkaMirrorMaker2Spec extends KafkaConnectSpec {
+public class KafkaMirrorMaker2Spec extends AbstractKafkaConnectSpec {
 
     private static final long serialVersionUID = 1L;
     
     private List<KafkaMirrorMaker2ClusterSpec> clusters;
     private String connectCluster;
     private List<KafkaMirrorMaker2MirrorSpec> mirrors;
-    private Map<String, Object> additionalProperties = new HashMap<>(0);
 
     @Description("Kafka clusters for mirroring.")
     public List<KafkaMirrorMaker2ClusterSpec> getClusters() {
@@ -61,21 +58,5 @@ public class KafkaMirrorMaker2Spec extends KafkaConnectSpec {
 
     public void setMirrors(List<KafkaMirrorMaker2MirrorSpec> mirrors) {
         this.mirrors = mirrors;
-    }
-
-    @Override
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @Override
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-    @Description("Bootstrap servers to connect to. This can be given as a comma separated list of _<hostname>_:\u200D_<port>_ pairs.")
-    @JsonProperty(required = false)
-    public String getBootstrapServers() {
-        return super.getBootstrapServers();
     }
 }
