+++ api/src/test/resources/io/strimzi/api/kafka/model/KafkaConnect.out.yaml
@@ -5,9 +5,10 @@ metadata:
   name: "test-kafka-connect"
 spec:
   replicas: 6
+  image: "foo"
+  bootstrapServers: "kafka:9092"
   config:
     name: "bar"
-  image: "foo"
   tolerations:
   - effect: "NoSchedule"
     key: "key1"
@@ -20,7 +21,6 @@ spec:
   logging:
     type: "inline"
     loggers: {}
-  bootstrapServers: "kafka:9092"
   externalConfiguration:
     env:
     - name: "SOME_VARIABLE"
