+++ api/src/test/resources/io/strimzi/api/kafka/model/KafkaMirrorMaker2.yaml
@@ -6,8 +6,6 @@ spec:
   image: foo
   replicas: 6
   connectCluster: target
-  config:
-    name: bar
   tolerations:
     - key: "key1"
       operator: "Equal"
