+++ helm-charts/strimzi-kafka-operator/templates/041-Crd-kafkaconnect.yaml
@@ -44,10 +44,12 @@ spec:
           properties:
             replicas:
               type: integer
-            config:
-              type: object
             image:
               type: string
+            bootstrapServers:
+              type: string
+            config:
+              type: object
             livenessProbe:
               type: object
               properties:
@@ -774,8 +776,6 @@ spec:
                   type: string
               required:
               - type
-            bootstrapServers:
-              type: string
             externalConfiguration:
               type: object
               properties:
