+++ helm-charts/strimzi-kafka-operator/templates/048-Crd-kafkamirrormaker2.yaml
@@ -43,8 +43,6 @@ spec:
               type: integer
             connectCluster:
               type: string
-            config:
-              type: object
             image:
               type: string
             livenessProbe:
@@ -853,97 +851,6 @@ spec:
                 required:
                 - sourceCluster
                 - targetCluster
-            authentication:
-              type: object
-              properties:
-                accessToken:
-                  type: object
-                  properties:
-                    key:
-                      type: string
-                    secretName:
-                      type: string
-                  required:
-                  - key
-                  - secretName
-                accessTokenIsJwt:
-                  type: boolean
-                certificateAndKey:
-                  type: object
-                  properties:
-                    certificate:
-                      type: string
-                    key:
-                      type: string
-                    secretName:
-                      type: string
-                  required:
-                  - certificate
-                  - key
-                  - secretName
-                clientId:
-                  type: string
-                clientSecret:
-                  type: object
-                  properties:
-                    key:
-                      type: string
-                    secretName:
-                      type: string
-                  required:
-                  - key
-                  - secretName
-                disableTlsHostnameVerification:
-                  type: boolean
-                maxTokenExpirySeconds:
-                  type: integer
-                passwordSecret:
-                  type: object
-                  properties:
-                    password:
-                      type: string
-                    secretName:
-                      type: string
-                  required:
-                  - password
-                  - secretName
-                refreshToken:
-                  type: object
-                  properties:
-                    key:
-                      type: string
-                    secretName:
-                      type: string
-                  required:
-                  - key
-                  - secretName
-                tlsTrustedCertificates:
-                  type: array
-                  items:
-                    type: object
-                    properties:
-                      certificate:
-                        type: string
-                      secretName:
-                        type: string
-                    required:
-                    - certificate
-                    - secretName
-                tokenEndpointUri:
-                  type: string
-                type:
-                  type: string
-                  enum:
-                  - tls
-                  - scram-sha-512
-                  - plain
-                  - oauth
-                username:
-                  type: string
-              required:
-              - type
-            bootstrapServers:
-              type: string
             externalConfiguration:
               type: object
               properties:
@@ -1034,21 +941,6 @@ spec:
                   type: object
                 requests:
                   type: object
-            tls:
-              type: object
-              properties:
-                trustedCertificates:
-                  type: array
-                  items:
-                    type: object
-                    properties:
-                      certificate:
-                        type: string
-                      secretName:
-                        type: string
-                    required:
-                    - certificate
-                    - secretName
             version:
               type: string
           required:
